cleanup! move proto-models code to own file
